+++ digest/CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## UNRELEASED
+### Added
+- `CustomizedInit` trait ([#1334]).
+
 ### Changed
 - `crypto-common` dependency bumped to v0.2 ([#1173])
 - Edition changed to 2021 and MSRV bumped to 1.57 ([#1173])